Adds BrokerHandler class, updates BrokerProcedures class. Moves backend methods to BrokerHandler.
